# Family-Nutrition-Planner/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NutriNext / ParivarSehat AI — Family Nutrition Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: Search and list recipes
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (name or ingredient)
        - name: cuisine
          in: query
          schema:
            type: string
          description: Filter by cuisine (e.g. "North Indian", "South Indian")
        - name: diet
          in: query
          schema:
            type: string
          description: Filter by diet type (e.g. "vegetarian", "vegan")
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category (e.g. "breakfast", "lunch", "dinner", "snack")
        - name: maxCalories
          in: query
          schema:
            type: integer
          description: Maximum calories
        - name: maxCookTime
          in: query
          schema:
            type: integer
          description: Maximum total cooking time in minutes
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated list of recipes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeListResponse"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get recipe details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Meal Plans ─────────────────────────────────────────────────────────────
...
  /meal-plans/{mealPlanId}/feedback:
    get:
      operationId: listMealFeedback
      tags: [meal-feedback]
      summary: Get all feedback for a meal plan
      parameters:
        - name: mealPlanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of feedback entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealFeedback"
    post:
      operationId: createMealFeedback
      tags: [meal-feedback]
      summary: Submit thumbs-up/down feedback for a meal
      parameters:
        - name: mealPlanId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealFeedbackBody"
      responses:
        "201":
          description: Created feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealFeedback"
...
  /nutrition-logs:
    get:
      operationId: listNutritionLogs
      tags: [nutrition-logs]
      summary: Get food intake logs for a family/member
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: integer
        - name: memberId
          in: query
          schema:
            type: integer
        - name: logDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of nutrition log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NutritionLog"
    post:
      operationId: createNutritionLog
      tags: [nutrition-logs]
      summary: Log a food intake entry (from scanner, manual, or AI)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNutritionLogBody"
      responses:
        "201":
          description: Created nutrition log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionLog"
  /nutrition-summary/{memberId}:
    get:
      operationId: getNutritionSummary
      tags: [nutrition-logs]
      summary: Get today's nutrition vs ICMR-NIN RDA targets for a member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nutrition summary with actual vs target breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionSummary"

  # ── Symptom Check ─────────────────────────────────────────────────────────
  /symptom-check:
    post:
      operationId: symptomCheck
      tags: [symptom-check]
      summary: AI nutritional advisor for reported symptoms (not medical diagnosis)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SymptomCheckBody"
      responses:
        "200":
          description: Nutritional insights and dietary suggestions for the reported symptoms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SymptomCheckResult"

  # ── Fasting Calendar ──────────────────────────────────────────────────────
  /fasting-calendar:
    get:
      operationId: getFastingCalendar
      tags: [fasting-calendar]
      summary: Get Indian festival/fasting days for a given month
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        "200":
          description: Fasting calendar with recommended foods per day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastingCalendar"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Family ────────────────────────────────────────────────────────────────
...
    Recipe:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameHindi:
          type: string
        cuisine:
          type: string
        category:
          type: string
          description: "breakfast, lunch, dinner, snack, dessert"
        diet:
          type: string
          description: "vegetarian, vegan, non-vegetarian"
        ingredients:
          type: string
        instructions:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        iron:
          type: number
        calcium:
          type: number
        vitaminC:
          type: number
        prepTimeMin:
          type: integer
        cookTimeMin:
          type: integer
        servings:
          type: integer
        costPerServing:
          type: number
          description: Estimated cost per serving in INR
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
      required:
        - id
        - name
        - cuisine
        - category
        - diet
        - calories
    RecipeListResponse:
      type: object
      properties:
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/Recipe"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - recipes
        - total
        - page
        - limit
        - totalPages

    # ── Meal Plans ─────────────────────────────────────────────────────────────
...
    MealFeedback:
      type: object
      properties:
        id:
          type: integer
        familyId:
          type: integer
        mealPlanId:
          type: integer
        dayIndex:
          type: integer
          minimum: 0
          maximum: 6
        mealType:
          type: string
        liked:
          type: boolean
        rating:
          type: integer
          minimum: 1
          maximum: 5
        skipReason:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - familyId
        - mealPlanId
        - dayIndex
        - mealType
        - liked
    CreateMealFeedbackBody:
      type: object
      properties:
        familyId:
          type: integer
        dayIndex:
          type: integer
          minimum: 0
          maximum: 6
        mealType:
          type: string
        liked:
          type: boolean
          default: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
        skipReason:
          type: string
        notes:
          type: string
      required:
        - familyId
        - dayIndex
        - mealType

    # ── Grocery Lists ─────────────────────────────────────────────────────────
...
    NutritionLog:
      type: object
      properties:
        id:
          type: integer
        familyId:
          type: integer
        memberId:
          type: integer
          nullable: true
        logDate:
          type: string
          format: date
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodDescription:
          type: string
          nullable: true
        calories:
          type: number
          nullable: true
        proteinG:
          type: number
          nullable: true
        carbsG:
          type: number
          nullable: true
        fatG:
          type: number
          nullable: true
        fiberG:
          type: number
          nullable: true
        ironMg:
          type: number
          nullable: true
        calciumMg:
          type: number
          nullable: true
        vitaminCMg:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        source:
          type: string
          enum: [manual, scanner, ai]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - familyId
        - logDate
        - mealType
    CreateNutritionLogBody:
      type: object
      properties:
        familyId:
          type: integer
        memberId:
          type: integer
        logDate:
          type: string
          format: date
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodDescription:
          type: string
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        fiberG:
          type: number
        ironMg:
          type: number
        calciumMg:
          type: number
        vitaminCMg:
          type: number
        imageUrl:
          type: string
        source:
          type: string
          enum: [manual, scanner, ai]
      required:
        - familyId
        - logDate
    NutrientValues:
      type: object
      properties:
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        iron:
          type: number
        calcium:
          type: number
        vitaminC:
          type: number
      required:
        - calories
        - protein
        - carbs
        - fat
        - fiber
        - iron
        - calcium
        - vitaminC
    NutritionSummary:
      type: object
      properties:
        member:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            age:
              type: integer
            gender:
              type: string
          required:
            - id
            - name
            - age
            - gender
        actual:
          $ref: "#/components/schemas/NutrientValues"
        target:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/NutrientValues"
        date:
          type: string
          format: date
      required:
        - member
        - actual
        - date

    # ── Symptom Check ─────────────────────────────────────────────────────────
    SymptomCheckBody:
      type: object
      properties:
        symptoms:
          type: array
          items:
            type: string
          minItems: 1
        age:
          type: integer
        gender:
          type: string
          enum: [male, female, other]
        existingConditions:
          type: array
          items:
            type: string
        language:
          type: string
          enum: [english, hindi]
      required:
        - symptoms
    SymptomCheckResult:
      type: object
      properties:
        disclaimer:
          type: string
          description: Required medical disclaimer
        nutritionalInsight:
          type: string
        dietarySuggestions:
          type: array
          items:
            type: string
        recommendedFoods:
          type: array
          items:
            type: string
        avoidFoods:
          type: array
          items:
            type: string
        seeDoctor:
          type: string
          description: When to seek medical attention
        urgency:
          type: string
          enum: [routine, soon, urgent]
      required:
        - disclaimer
        - nutritionalInsight
        - dietarySuggestions
        - recommendedFoods
        - avoidFoods
        - seeDoctor
        - urgency

    # ── Fasting Calendar ──────────────────────────────────────────────────────
    FastingDay:
      type: object
      properties:
        date:
          type: string
          format: date
        name:
          type: string
        nameHindi:
          type: string
        fastingType:
          type: string
          enum: [full, partial, none]
        recommendedFoods:
          type: array
          items:
            type: string
      required:
        - date
        - name
        - nameHindi
        - fastingType
        - recommendedFoods
    FastingCalendar:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        fastingDays:
          type: array
          items:
            $ref: "#/components/schemas/FastingDay"
        note:
          type: string
      required:
        - year
        - month
        - fastingDays
        - note

    # ── Voice Profile Parsing ──────────────────────────────────────────────────
...