# Family-Nutrition-Planner/lib/api-zod/src/generated/types/recipe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriNext / ParivarSehat AI — Family Nutrition Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Recipe {
  id: number;
  name: string;
  nameHindi?: string;
  cuisine: string;
  /** breakfast, lunch, dinner, snack, dessert */
  category: string;
  /** vegetarian, vegan, non-vegetarian */
  diet: string;
  ingredients?: string;
  instructions?: string;
  calories: number;
  protein?: number;
  carbs?: number;
  fat?: number;
  fiber?: number;
  iron?: number;
  calcium?: number;
  vitaminC?: number;
  prepTimeMin?: number;
  cookTimeMin?: number;
  servings?: number;
  /** Estimated cost per serving in INR */
  costPerServing?: number;
  tags?: string[];
  imageUrl?: string;
}