# Family-Nutrition-Planner/lib/api-zod/src/generated/types/symptomCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriNext / ParivarSehat AI — Family Nutrition Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { SymptomCheckResultUrgency } from "./symptomCheckResultUrgency";

export interface SymptomCheckResult {
  /** Required medical disclaimer */
  disclaimer: string;
  nutritionalInsight: string;
  dietarySuggestions: string[];
  recommendedFoods: string[];
  avoidFoods: string[];
  /** When to seek medical attention */
  seeDoctor: string;
  urgency: SymptomCheckResultUrgency;
}